agregando DTOs a producto y categoria

+++ messages/src/main/java/com/coin/messages/service/CategoriaService.java
@@ -1,5 +1,7 @@
 package com.coin.messages.service;
 
+import com.coin.messages.dto.request.CategoriaRequest;
+import com.coin.messages.dto.response.CategoriaResponse;
 import com.coin.messages.model.Categoria;
 import com.coin.messages.repository.CategoriaRepository;
 import org.springframework.stereotype.Service;
@@ -14,22 +16,41 @@ public class CategoriaService {
         this.repository=repository;
     }
 
-    public List<Categoria> listar(){
-        return repository.findAll();
+    public List<CategoriaResponse> listar(){
+        return repository.findAll().stream().map(this::convertirResponse).toList();
     }
-    public Categoria obtener(Long id){
-        return repository.findById(id).orElseThrow(()->new RuntimeException("No se encontro la categoria con id:"+id) );
+    public CategoriaResponse obtener(Long id){
+        Categoria categoria = repository.findById(id).orElseThrow(()->new RuntimeException("No se encontro la categoria con id:"+id) );
+        return convertirResponse(categoria);
     }
-    public Categoria guardar(Categoria categoria){
-        return repository.save(categoria);
+    public CategoriaResponse guardar(CategoriaRequest categoriaRequest){
+        Categoria categoria  = new Categoria();
+        categoria.setNombre(categoriaRequest.getNombre());
+        categoria.setDescripcion(categoriaRequest.getDescripcion());
+        Categoria guardar = repository.save(categoria);
+
+        return this.convertirResponse(guardar);
     }
-    public Categoria actualziar(Long id, Categoria categoria){
+    public CategoriaResponse actualziar(Long id, CategoriaRequest categoriaRequest){
         Categoria existe = repository.findById(id).orElseThrow(()->new RuntimeException("No existe la categoria"+id) );
-        existe.setNombre(categoria.getNombre());
-        existe.setDescripcion(categoria.getDescripcion());
-        return repository.save(existe);
+        existe.setNombre(categoriaRequest.getNombre());
+        existe.setDescripcion(categoriaRequest.getDescripcion());
+        Categoria actualziar = repository.save(existe);
+        return this.convertirResponse(actualziar);
     }
     public void eliminar(Long id) {
+        if (!repository.existsById(id)) {
+            throw  new RuntimeException("No existe el categoria con id:"+id);
+        }
         repository.deleteById(id);
+    }
+
+    //ES OTRA MANERA DE HACER EL CASTEO DE TIPOS PERO ES MEJOR UNAR UN MAPER
+    private CategoriaResponse convertirResponse(Categoria categoria){
+        return new CategoriaResponse(
+                categoria.getId(),
+                categoria.getNombre(),
+                categoria.getDescripcion()
+        );
     }
 }
