+++ messages/src/main/java/com/coin/messages/service/ProductoService.java
@@ -1,10 +1,13 @@
 package com.coin.messages.service;
 
+import com.coin.messages.dto.request.ProductoRequest;
+import com.coin.messages.dto.response.ProductoResponse;
 import com.coin.messages.model.Categoria;
 import com.coin.messages.model.Producto;
 import com.coin.messages.repository.CategoriaRepository;
 import com.coin.messages.repository.ProductoRepository;
 import org.springframework.stereotype.Service;
+import com.coin.messages.mapper.ProductoMapper;
 
 import java.util.List;
 
@@ -13,41 +16,44 @@ public class ProductoService {
 
     private final ProductoRepository repo;
     private final CategoriaRepository repoCategoria;
+    private final ProductoMapper mapper;
 
-    public ProductoService(ProductoRepository repo, CategoriaRepository repoCategoria) {
+    public ProductoService(ProductoRepository repo, CategoriaRepository repoCategoria, ProductoMapper mapper) {
         this.repo = repo;
         this.repoCategoria=repoCategoria;
+        this.mapper=mapper;
     }
 
-    public List<Producto> listar() {
-        return repo.findAll();
+    public List<ProductoResponse> listar() {
+        return repo.findAll().stream().map(mapper::toProductoResponse).toList();
     }
 
-    public Producto obtener(Long id){
-        return repo.findById(id).orElseThrow(()-> new RuntimeException("Producto no encontrado con id:"+id));
+    public ProductoResponse obtener(Long id){
+        Producto producto = repo.findById(id).orElseThrow(()-> new RuntimeException("Producto no encontrado con id:"+id));
+        return  mapper.toProductoResponse(producto);
     }
 
-    public Producto actualizar(Long id, Producto producto) {
+    public ProductoResponse actualizar(Long id, ProductoRequest productoRequest) {
         Producto exitente = repo.findById(id).orElseThrow(()-> new RuntimeException("Producto no encontrado con id:"+id));
+        Categoria categoria = repoCategoria.findById(productoRequest.getCategoriaId()).orElseThrow(()-> new RuntimeException("Categoria no encontrado con id:"+productoRequest.getCategoriaId()));
+        mapper.updateProducto(exitente, productoRequest, categoria);
 
-        Categoria categoria = repoCategoria.findById(producto.getCategoria().getId()).orElseThrow(()-> new RuntimeException("Categoria no encontrado con id:"+producto.getCategoria().getId()));
-        producto.setCategoria(categoria);
-
-        exitente.setNombre(producto.getNombre());
-        exitente.setDescripcion(producto.getDescripcion());
-        exitente.setCategoria(producto.getCategoria());
-        exitente.setPrecio(producto.getPrecio());
-        exitente.setStock(producto.getStock());
-        exitente.setImagenUrl(producto.getImagenUrl());
-        return  repo.save(exitente);
+        Producto actualizado = repo.save(exitente);
+        return  mapper.toProductoResponse(actualizado);
     }
-    public Producto guardar(Producto p) {
-        Categoria categoria = repoCategoria.findById(p.getCategoria().getId()).orElseThrow(()-> new RuntimeException("Categoria no encontrado con id:"+p.getCategoria().getId()));
-        p.setCategoria(categoria);
-        return repo.save(p);
+    public ProductoResponse guardar(ProductoRequest p) {
+        Categoria categoria = repoCategoria.findById(p.getCategoriaId()).orElseThrow(()-> new RuntimeException("Categoria no encontrado con id:"+p.getCategoriaId()));
+        Producto producto = new Producto();
+        mapper.updateProducto(producto,p,categoria);
+        Producto guardado = repo.save(producto);
+
+        return mapper.toProductoResponse(guardado);
     }
 
     public void eliminar(Long id) {
+        if (!repo.existsById(id)) {
+            throw new RuntimeException("Producto no encontrado");
+        }
         repo.deleteById(id);
     }
 }
